move api type definition to apiTypes.ts
